Move `HostHandle` to separate file
